requirements.txt
torch~=2.1.0
mlflow==2.8.0
torchvision~=0.16.0
numpy~=1.26.1
lightning~=2.1.0
torchmetrics~=1.2.0
umap-learn
seaborn~=0.13.0
pandas~=2.1.2
matplotlib~=3.8.1
omegaconf~=2.3.0
hydra-core~=1.3.2